Use data() over c_str() where null termination is not required
For preperation to take string_view to use use data() over c_str(). It
is certain that this commit alone will not break anything but we have to
keep in mind that now these pleces are marked like so that no null
termination is needed in data and in future that can happend.