Use std::string straight without .c_str()
In many places we can use string straight witout using .c_str(). There
is benefit of little bit cleaner code and also code does not have to
take strlen() anymore.

Note that these where hand selected because not all can be just changed.
Sometimes ugly hack with null termination might be used. Example if
string is "aa\0aa" and we take .c_str() and passed this to function
which needs string this would mean it get's just "aa\0". That is ugly
hack and should always be documented, but some code seems that it might
rely on that behavier.